Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> Oh, did you need the casts because you’re coming out of an enum _class_? Maybe it should just be “enum : unsigned int”.

Yeah, it's because they're enum class's. I like the namespacing of enum class, and I like the ability to specify the underlying type, but I'm growing to dislike the lack of implicit conversion to the underlying type.

So I can either keep the casts, or the names have to be something like kTerminationCodeNoCrashDump, kTerminationCodeSnapshotFailed, etc. I'll go with that in this case if you don't have a preference.

Patch-set: 4
Label: Code-Review=0