Annotations replacement prototype
Based on idea in
https://groups.google.com/a/chromium.org/d/msg/crashpad-dev/1x7W6-WXPKs/3adreFkiBgAJ.

Confirmed that it compiles to something reasonable on Mac too since that
post.

Unfortunately, I forgot that it's not possible to extern something in a
different namespace, so rather than simply having a DECLARE() in .cc and
a SET() as necessary, the files that do SET() also need a DECLARE() at
file scope to extern the global variable.

R=mark@chromium.org
BUG=

Change-Id: I652c0d5e538b5880e818c1acecc30d66e8c29472