win: Start crashpad_handler by inheriting connection data to it
Previously, StartHandler() launched the handler process, then connected
over a pipe to register for crash handling. Instead, the initial client
can create and inherit handles to the handler and pass those handle
values and other data (addresses, etc.) on the command line.

This should improve startup time as there's no need to synchronize with
the process at startup, and allows avoiding a call to CreateProcess()
directly in StartHandler(), which is important for registration for
crash reporting from DllMain().

Note: API change; UseHandler() is gone.

XXX update docs

XXX Mac needs to be made to match on its startup paths

BUG=567850,656800

Change-Id: I1602724183cb107f805f109674c53e95841b24fd