Update patch set 3
Patch Set 3:

> A Sleep() seems "ok, but still one more thing" I guess. We can only signal
> until the end of StartHandlerProcess(), without making it more complicated
> and/or WaitForHandlerStart() non-optional. I guess if CreateProcess()
> succeeded (the end of StartHandlerProcess()) then arguably either all is well,
> or we're moving on to unexpected termination, rather than a failure to start.

I wasn’t thinking that we had to mandate WaitForHandlerStart(), but if we reached the end of StartHandlerProcess() (which maybe also ultimately does the WaitNamedPipe() like you suggested previously, but you might be saving that for the third follow-up), that’s a good sign that things have been set up. Whether or not the caller cares to check in on it is distinct.

> I'm mostly just nervous about putting a potentially infinite loop at the head
> of the exception function. I guess we can hope UMA will tell us if things are
> going terribly wrong.

I hear ya. But on the other side of it, I’m leery of barreling into the exception handler in a questionable state. We wouldn’t loop unless things look like they’re in the process of being set up. We can also reason that an accidental infinite spin there is way less likely than hitting the Sleep(60) in a success-impossible situation, and from a user’s perspective, I don’t know that there’s that much difference between a Sleep(60) and an infinite loop anyway.

Patch-set: 3
Label: Code-Review=0