Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > A Sleep() seems "ok, but still one more thing" I guess. We can only signal
> > until the end of StartHandlerProcess(), without making it more complicated
> > and/or WaitForHandlerStart() non-optional. I guess if CreateProcess()
> > succeeded (the end of StartHandlerProcess()) then arguably either all is well,
> > or we're moving on to unexpected termination, rather than a failure to start.
> 
> I wasn’t thinking that we had to mandate WaitForHandlerStart(), but if we reached the end of StartHandlerProcess() (which maybe also ultimately does the WaitNamedPipe() like you suggested previously, but you might be saving that for the third follow-up), that’s a good sign that things have been set up. Whether or not the caller cares to check in on it is distinct.

A simple WaitNamedPipe() doesn't end up being sufficient, because if the server side of the pipe *exists* then there's an instance available from WaitNamedPipe()'s point of view. My mistake in previous comments was that I was sort of thinking that the handle would disappear immediately when we close the client's copy, but it doesn't -- the OS has to actually get around to really closing down the pipe once the last process has stopped using it.

That is,

pipe.reset();
Sleep(100);
WaitNamedPipe(pipe_name, ...);

"works", but the Sleep() is required.

So, instead, I added an empty "ping" message to the communication protocol. The initial client actually connects, sends an empty message, and receives an empty response, so we can be confident the server is up and ready to go.

(With this change, I'm starting to wonder about the utility having the asynchronous_start flag at all, as the behaviour is pretty reasonable in async mode, even if the client doesn't really care about the way in which the handler is started.)

> 
> > I'm mostly just nervous about putting a potentially infinite loop at the head
> > of the exception function. I guess we can hope UMA will tell us if things are
> > going terribly wrong.
> 
> I hear ya. But on the other side of it, I’m leery of barreling into the exception handler in a questionable state. We wouldn’t loop unless things look like they’re in the process of being set up. We can also reason that an accidental infinite spin there is way less likely than hitting the Sleep(60) in a success-impossible situation, and from a user’s perspective, I don’t know that there’s that much difference between a Sleep(60) and an infinite loop anyway.

Yup, that makes sense.

My only concern is that startup is getting hairy, but I guess we signed up for that with --initial-client-data. (Or maybe it was farther back when we started writing a crash handler.)

Patch-set: 3
Label: Code-Review=0