NOT FOR CHECKIN: Review helper for AsciiDoc→Markdown conversion
Don’t check this in. Use
https://chromium-review.googlesource.com/408256/ instead.

This exists only as a review helper. Enough of the file content in the
real version moved around rendering git unable to follow history. This
version has all of the *.ad→*.md file content changes from the other
one, but leaves everything in the un-renamed *.ad files. This version
doesn’t have any of the non-*.ad/*.md changes.

BUG=crashpad:138

Change-Id: I7d8474ca507cf846dbef75c381b7d9b6a865dd7e