Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > Looks like all of the Crashpad bots are down-rev on the SDK too.
> > 
> > Would you prefer that I resurrect the fallback definitions, or should we update the SDK on those bots?
> 
> We might as well just update them. I'll see if I still have access.

Still working on it. The SDK won't install because it wants an update to the universal CRT. *That* update won't install because Windows Update hangs. Apparently there's a Windows Update Client Update which will fix that, but of course that's not that easy to install because, you guessed it, it's updated by Windows Update.

Soon I will win. Maybe.

Patch-set: 4