Update patch set 4
Patch Set 4:

> +Bruce -- One minor reason to update to SDK 14393. I don't know if
 > there's any other outstanding benefits or drawbacks to doing so
 > though. The only reference I see now is part of https://crbug.com/427616.

I created a VS 2015 Update 3 package with 10.0.14393 when I created the package that we are currently using. I meant to upload it (and make it a selectable option) but apparently never did.

So, I just uploaded the CL with the magic hash to https://codereview.chromium.org/2492573002. I can land it as-is to allow testing, or we could change to that hash.

I don't know what the deal is with the UCRT updates. We don't ship with the UCRT so we are insulated to some extent from whatever the consequences are.

Patch-set: 4