Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> It may be that Windows 7 is not properly supported anymore. The 10.0.14393 version of WPT (Windows Performance Toolkit) will install on Windows 7 but won't run, which means your machine ends up in a partially broken state if you install it.
> 
> We may need to move our build machines to Windows 10 at some point. I inquired about this last month (due to a svchost.exe memory leak on Windows 7 that our builders hit occasionally) but for the moment we are content with just rebooting our build machines more frequently. These issues may give us another reason to start switching (uggh).

Yeah, it's a bit of a mess. It's definitely not anything we need for the build that's trying to install, so we'd be fine with the packaged toolchain.

Unfortunately these are client bots, not chrome bots, so swarming, etc. is more challenging script-wise (not to mention it adds a lot of complexity when we just want to run some small unittests). But we need to test on Win7, so these bots kind of have to be Win7, or we have to write custom goop to handle a builder/runner split.

Similarly, we need to run on an actual x86 machine (not just wow64), which is also a pain for similar sorts of reasons (no ninja x86 binary, VS 2015 might not like x86 Win7, ...)

Someday maybe we'll just be able to sync clang and an SDK repo. But that day is not today.

Patch-set: 4