Update patch set 1
Patch Set 1:

> Do you know if this is a violation of the standard or is GCC buggy?

Name hiding in general is allowed (C++11 §3.3.10), but here, the declaration of the function uses the pre-hidden name, a type, before replacing it with a function. C++11 §3.3.7(1)(2) calls me an outlaw because the hidden name was used to declare the class. But it also says “no diagnostic is required,” so Clang and MSVC aren’t off base for remaining silent on the issue either. When everybody wins, nobody wins!

(Ask me if you don’t have a copy of the standard but want to read more.)

Patch-set: 1
Label: Code-Review=0