Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> OK. But I’m no fool, I know that signals aren’t a native Windows thing. So this is some CRT emulation? And if some module is using its own static CRT, we’re powerless to catch crashes that it interprets as signals?
> 
> I’m really hopeful that bug 133 will give us a no-fuss catch-all when it works.

Yeah, the raise() is implemented in the CRT, and it's always synchronous. So this is really just a way of hooking into a code path that gives us access to abort() to catch it before it does the __fastfail().

I don't think we have a static-CRT problem as it's really just the plain old C implementation of abort() that we'd miss out on if it was using a different/custom CRT. The OS-level thing is still the unhandled exception filter, and we're installing that as enthusiastically as we can (at least in Chrome).

Patch-set: 2