Update patch set 2
Patch Set 2:

> Change has been successfully cherry-picked as 68095b6a4e6148125fe631c63469496d2afca29c by Scott Graham

Windows was still broken after this because of the mucking around I did when I was trying to get the new SDK installed (I broke the debugging tools installation).

That's repaired now, and all bots are green for the first time in quite a while, as the Microsoft symbol server seems to be alive again.

Patch-set: 2