win: Attempt to fix unloaded modules list by using RtlGetUnloadEventTraceEx
I haven't been able to reproduce this locally, but we see it in crash
dumps. My assumption is that RTL_UNLOAD_EVENT_TRACE isn't correct for
some OS levels. Instead of trying to finess and test that, use
RtlGetUnloadEventTraceEx() instead of RtlGetUnloadEventTrace(), which
returns an element size. (This function is Vista+ which is why it wasn't
used the first time around.)

R=mark@chromium.org
BUG=chromium:620175

Change-Id: I4d7080a03623276f9c1c038d6e7329af70e4a64c