Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> LGTM.
> 
> I don’t know if size confusion was really the problem, because you’d expect at least the first unloaded module to be somewhat reasonable in that case. Although perhaps the first one didn’t have a name set causing us to omit it from the dump without leaving any evidence behind?

Yeah. I'm wondering now if it was actually that I misinterpreted the documentation for RtlGetUnloadEventTrace(), and it returns a pointer to a pointer too (which the Ex() version does). But the tests pass locally on Win10, and on Win7 bots, and they unload 2 modules. So I'm confused. I'll land this, but do a bit more testing anyway.

Patch-set: 3
Label: Code-Review=0