Update patch set 2
Patch Set 2:

https://chromium.googlesource.com/crashpad/crashpad/+log/02dcb4361f6ce2a6bd37db048fcdb856b28923fe tells me that you’re based on 777634b1ebad7149d0af41686f1f36126b4ea984 (2016-11-23), which is pretty out of date compared to the current master, 0567536f86fb10f9663fb30d6ebf08a7c35b975d (2016-12-17). DEPS has changed in that window (f66d5df30cb85a637c197bbab1541367f92f3c62) which is why Gerrit won’t let you submit.

A local rebase will do the trick. If it can be automatically rebased conflict-free, the rebase button in the Gerrit UI should also take care of it, and that should enable the Submit button.

I think that this is kind of unclear too, but I guess I just learned it as a Gerrit quirk that we have to deal with. We could file a bug report and ask that Submit be made to do Rebase+Submit for things that can be automatically rebased without conflict.

Patch-set: 2