Update comments in IsExceptionNonfatalResource() given 10.12 source
With reference to 10.12 source, commentary regarding RESOURCE_TYPE_IO
can be authoritative.

Cursory examination of 10.12 source reveals that RESOURCE_TYPE_MEMORY
can now be fatal, although deeper examination reveals that this is
impossible on macOS. State this authoritatively as well. (It’s likely
that in iOS 10, the iOS “Jetsam” out-of-memory killer’s actual killing
functionality moved into the kernel rather than being performed by an
out-of-kernel EXC_RESOURCE exception handler. It was probably always
possible for these exceptions to be fatal on iOS.)

BUG=crashpad:124

Change-Id: I52124c68fe017015983ab46e54006ba97ecd0142