Update patch set 3
Patch Set 3:

This never got checked in, which may have been because the SDKs weren’t all up to snuff. But I think that the SDKs should be in good shape now, so I’ll check this in later today unless you can think of a reason not to.

Patch-set: 3
Label: Code-Review=0