Update patch set 16
Patch Set 16:

> Patch Set 16:
> 
> Hey Mark, do you think this approach is too unpleasant to consider? I haven't implemented the reading test for Mac yet, but I don't think it should be too complicated.
> 
> I'm not overly set on using this particular approach, but I feel we should do something one way or the other because the Chrome situation is really a mess.

I've been kicking around an idea that's similar but slightly different and doesn't involve a macro. I agree the situation in Chrome is now terrible.

My idea was to have a crash key and its storage declared in the same file as its user. But rather than using a macro to get it in a special section, have the Set() methods on the CrashKey object add a pointer to the object to a global, lock-free linked list. This approach has two benefits over this approach:

1) No macros to place the data in a special section
2) The same code could technically be used in Chromium to back the Linux implementation of crash keys until we get Crashpad working there

I kinda wanted to pick this up after perf, but people seem to be making a lot of noise *right now*.

Patch-set: 16
CC: Robert Sesek <1144624@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>