Update patch set 1
Patch Set 1: Code-Review-1

This was intentionally not implemented in Crashpad. Magic HTTP form parameters are on the way out. Everything should be encoded in the minidump file. The process uptime is known as the difference between MINIDUMP_MISC_INFO::ProcessCreateTime and MINIDUMP_HEADER::TimeDateStamp. Our server-side infrastructure knows how to compute the process uptime in this way without having to do the identical computation on the client side and transmitting it as “ptime.” Your server-side infrastructure could easily do the same.

Because the minidump format only gives one-second-resolution time_t values for these fields, I would support adding something to the extension structure that conveys higher-resolution (microsecond or nanosecond) values. But this would be an extremely low priority.

This patch is wrong for a number of other reasons. It calculates “ptime” as the difference between the handler process start time and the crash process exception time, but the handler start time has nothing to do with the crash process start time. It is often earlier than, but can also be later than, the crash process start time. It also neglects the fact that we already capture the exception time in a better location and make it available via Snapshot::SnapshotTime(), and we also capture the actual crash process start time and make that available via Snapshot::ProcessStartTime().

Patch-set: 1
Reviewer: Mark Mentovai <1174099@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-1