Update patch set 17
Patch Set 17:

(4 comments)

I was excited to take this for a spin, but I couldn’t actually build it in any configuration except arm64 with Android API 21.

Per https://chromium.googlesource.com/crashpad/crashpad/+/master/doc/developing.md#Building-Android, we need to care about arm/API16, arm64/API21, and x86/API16, x86_64/API21. Robert warned me that Chrome still also needs to build on mips and mips64 as well. We should also be able to build well on a contemporary Linux (I tested with Ubuntu 16.10). This can get tricky as especially the older NDKs often have broken or unusable headers. We’ll need to provide some compat definitions (take a look in the compat directory).

I think that for now, so as to avoid holding this patch up any further, we should proceed as-is without worrying about fixing the NDK/API-related problems. We should save the fixes for those for a follow-up (it quite possibly won’t need to touch any of this code anyway, if we do it all in compat). That leaves just the user_pt_regs problem from line 309.

I believe that the correct types to use on x86 and x86_64 are user_regs_struct, and on 32-bit arm, pt_regs. But please test this! Once we’re sure of the right types, we can work them into:

#if defined(ARCH_CPU_X86_FAMILY)
  using PrStatusType = user_regs_struct;
#elif defined(ARCH_CPU_ARMEL)
  using PrStatusType = pt_regs;
#elif defiend(ARCH_CPU_ARM64)
  using PrStatusType = user_pt_regs;
#endif

Patch-set: 17
Reviewer: Mark Mentovai <1174099@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>