linux: Lazily initialize ProcessInfo’s Is64Bit() and StartTime()
Lazy initialization is particularly beneficial for Is64Bit(), which uses
a different (ptrace()-based) approach than the rest of the class (which
is /proc-based). It is possible for the /proc-based Initialize() to
succeed while ptrace() would fail, as it typically would in the
ProcessInfo.Pid1 test. Because this test does not call Is64Bit(),
permission to ptrace() shouldn’t be necessary, and in fact ptrace()
shouldn’t even be called.

In order to make ProcessInfo::StartTime() operate lazily and cache its
result, ProcessInfo::ProcessStartTime() must be made non-const. const
removal bubbles through the codebase a bit. See bug crashpad:9.

Bug: crashpad:9, crashpad:30
Test: crashpad_util_test ProcessInfo.*