util/file: Handle oversized reads and writes gracefully
file_io and the FileReader family had a few loose ends regarding big
reads and writes. It’s not likely that we’ve experienced these
conditions yet, but they’d be likely to appear in a potential future
involving full memory dumps. This specifies the behavior with large
reads and writes, consolidates some logic, and improves some interfaces.

ReadFile() should always return after a short read, and in fact does
return after short reads since 00b64427523b. It is straightforward to
limit the maximum read size based on a parameter limitation of the
underlying operation, or a limitation of the type used for
FileOperationResult.

In contrast, WriteFile() should always retry after a short write,
including a write shortened because of a parameter limitation of the
underlying operation. This allows its return value to be simplified to a
“bool”.

The platform-specific WriteFile() code has been moved to
internal::WriteFileInternal(), and WriteFile() is a platform-independent
loop that retries following a short write.

The platform-agnostic ReadFileExactlyInternal() implementation has been
refactored into internal::ReadExactlyInternal so that it can be used by
FileReaderInterface::ReadExactly(), which had a nearly identical
implementation.

Test: *
Change-Id: I487450322ab049c6f2acd4061ea814037cc9a864