Update patch set 1
Patch Set 1:

Hey Mark,

this is a strawman CL for how to shoehorn the user stream extensibility call-outs into the handler. I'm open to suggestions on where to declare the interface, better ways to fit this in, naming - you name it.
I guess to test this I'll have to instantiate the handler with the new main function and tickle it...

Siggi

Patch-set: 1
Reviewer: Sigurður Ásgeirsson <1177549@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>