Update patch set 16
Patch Set 16:

OK, I took yet another close look at what I was seeing a few days ago that caused the need for the cookie thing. It wasn’t extra threads running in our gtest process. Robert seems to think that we can be about as assured that we’re the only thread in a gtest-based test on Android as we can on other platforms. That is to say, it’s probably safe enough to assume that this is the case here as to not worry about it until proven otherwise.

My EXPECT_DEATH()-based tests for ScopedMmap weren’t having other mapping show up on top of the existing mapped region because some other thread was grabbing them. It was happening because EXPECT_DEATH()-based tests are fork()/clone()-based. That’s what was opening the window for lots of other code to run. See https://github.com/google/googletest/blob/master/googletest/docs/AdvancedGuide.md#how-it-works for a description of how death tests work.

I’ll send out a change later tonight to update the comments in the ScopedMmap test to not be so misleading about threadiness. In the mean time, for your change right here, it sounds like we fortunately can proceed with these “unmapped” tests, and we can in fact expect that a region we unmap will remain unmapped until we try to read it, and that the read should fail.

I still want to avoid the double-munmap(), and I think that the unmapped test can be simplified to just use single ScopedMmap that you Reset() without having to deal with unmapping single pages any more granularly than that. If you want to check the return value of munmap(), you can modify Reset() as discussed earlier. But I’m not being prescriptive about the single-region thing here: if you have other reasons to want to be more granular and unmap things separately that aren’t obvious, we can talk about it too.

Portions of the unmapped tests in the Mach task_memory_test.cc do look suspect now that I’m referring to them again, so I think that there may be room for simplification there too.

Patch-set: 16
CC: Robert Sesek <1144624@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Mark Mentovai <1174099@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>